// packages/generator/src/overwrite-strategy.d.ts
export declare enum OverwriteStrategy {
    Never = "never",
    Ask = "ask",
    Overwrite = "overwrite",
    MergeLater = "merge-later"
}
export declare enum NamingStrategy {
    AsIs = "as-is",
    Kebab = "kebab",
    Pascal = "pascal",
    Camel = "camel",
    Snake = "snake"
}
//# sourceMappingURL=overwrite-strategy.d.ts.map